+++ NEWS.md
@@ -1,6 +1,7 @@
 # scorecard 0.1.6
 
 * Fixed a bug in woebin_adj function when all_var == FALSE and the breaks of all variables are perfect. 
+* Provide parallel computation (foreach with parallel backend) in the functions of woebin and woebin_ply.
 
 # scorecard 0.1.5